+++ changelog.md
@@ -1,6 +1,6 @@
 ## Forestry 1.0.15
 - Fix inconsistent ordering of life stages in API
-- 
+- Fix Snowing effect particles and server crash
 
 ## Forestry 1.0.14
 - Add Falkory's textures (#33)